requirements.txt
boto3==1.15.1
botocore==1.18.1
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2
django-autoslug==1.9.7
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-storages==1.10.1
gunicorn==20.0.4
idna==2.9
jmespath==0.10.0
Pillow==7.1.2
psycopg2==2.8.2
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.2.0
djangorestframework==3.12.2
django-filter==2.4.0
Markdown==3.3.3